removed a few non-state functions.  Items related to the User Interface now live in Output pkg.  Removed a redundant showState() func.
