+++ output/simpleui.go
@@ -1,0 +1,18 @@
+package output
+
+import (
+	"encoding/json"
+
+	"github.com/mdsauce/sauced/logger"
+	"github.com/mdsauce/sauced/manager"
+)
+
+// ShowStateJSON will pretty print JSON
+func ShowStateJSON() {
+	last := manager.GetLastKnownState()
+	lastJSON, err := json.MarshalIndent(last, "", "    ")
+	if err != nil {
+		logger.Disklog.Warnf("Could not format JSON with Indents: %v", err)
+	}
+	logger.Disklog.Info(string(lastJSON))
+}
